python/test/requirements.txt
# Allow any version of six
six
# Specify a particular version (not the latest) of attrs.
attrs==18.1.0
